moving test to state as requested, had to duplicate boiler plate code for creating core.Service instance that could be refactored further or moved into utils
